[REFACTORING] - Skiller ut org.enhet
Etablerer egen pakke for org.enhet. Der er det et lite subdomene, og en gateway for å snakke med ekstern tjeneste.
Nå er det litt sirkulære avhengigheter - dette fikses senere ved å innføre et interface i domenepakken, slik at avhengigheten kan gå inn mot domenet, og ikke motsatt vei.
