+++ src/main/java/no/nav/fo/veilarbregistrering/domain/Veileder.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import lombok.experimental.Accessors;
+import no.nav.fo.veilarbregistrering.orgenhet.NavEnhet;
 
 @Data
 @Accessors(chain = true)
