+++ src/main/java/no/nav/fo/veilarbregistrering/orgenhet/NavEnhet.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.domain;
+package no.nav.fo.veilarbregistrering.orgenhet;
 
 import lombok.*;
 import lombok.experimental.Accessors;
