+++ src/test/java/no/nav/fo/veilarbregistrering/service/ManuellRegistreringServiceTest.java
@@ -2,6 +2,7 @@ package no.nav.fo.veilarbregistrering.service;
 
 import no.nav.dialogarena.aktor.AktorService;
 import no.nav.fo.veilarbregistrering.db.ArbeidssokerregistreringRepository;
+import no.nav.fo.veilarbregistrering.orgenhet.EnhetOppslagService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
